tokenizer regexp speedups
This commit adds a few simple changes to speedup
the tokenizing process:
- subn()[0] -> sub() (avoids collecting a return value that is ignored)
- r'[a-f]', re.I -> r'[A-Fa-f]' (avoid case insensitive matching which
is expensive)
- '(foo|bar)' -> 'foo|bar' (avoid grouping when not referenced)
- '(...)" -> '(?:....)' (avoid capturing grouping)

It also replaces the IPV6_RE with a slightly less expensive regexp.
this one isn't pixel perfect, but a lot faster to match.

Overall with a sample input a logreduce dir-run improved from
from 0.375MB/s to 0.687MB/s.

Change-Id: I0d6b523ce3452faf356597b61dfb4d70105674ab