Fix overly greedy date tokenization
The tokenization process was overly eager in matching a date
string (like a month like "jan") within hashes, which broke
subsequent other tokenization attempts. Reorder the matching
to ensure that the "longest" / most specific regex is matched
first and tighten the regexp to only match on word boundaries.

Add tests to avoid regressing this again.

Change-Id: I9e688e65d3e7e61e2190f17dcbdcc25565cdb547