Vectorizer optimisation: don't do word analysing
The way the Tokenizer.process() works is that it already does the
word splitting and substitution like we want, so the Vectorizer can
save some time and simply use str.split() to get the words, rather
than trying to regexp it once more.

This improves performance of vectorization from 296.716MB/s
to 561.717MB/s on my machine.

Change-Id: Ibf423e5a53883dac3e20836f94a1ce5ed44e8f18