Don't store duplicate data in model
As we're only running for k-NN with k=1, all we're effectively
doing is a fuzzy search of the text line under test in the
data model, and as we're only are getting max one distance back,
storing the same data duplicated makes no sense other than
to burn cpu.

This keeps the test time on models created from more than
one training input much more consistent. Before the model
was growing linearly with the number of jobs included,
which caused testing to grow by roughly 2s per job inclusion.

With this change, test time of a default 2 job model
decreases from 10s to 6.7s. With a 64 job model the testing
time reduces from 154s to 7.3s, providing 2200% speedup.

Change-Id: I6810a8db1cf833e1fdc36b0564d854fd47be0d33