tests: Use slices instead of `vec!`
Here we have constant data; there's no need to heap-allocate a
`Vec` to hold it.

Performance is obviously irrelevant for the tests here,
but it's useful to use slices where we can because it can happen
in more production code paths.

However, what you may have gotten tripped up on here is:
https://doc.rust-lang.org/nightly/edition-guide/rust-2021/IntoIterator-for-arrays.html

Basically in Rust 2018 and below, `["foo", "bar"].iter()` gives you `&&str`
by default.  So for now let's use the suggested workaround.

Or, you could bump to 2021 edition.